feat: added multi-char key detection
Some keys send multiple characters to the terminal. For example, left
arrow in the kitty terminal sends \EOD, where \E is escape. In order
for the terminal session to correctly interpret this, the characters
apparently need to be sent together. Previously, each character in the
script file was sent separately.

Multi-character keys in the script are now detected and sent together.
However, the issue then becomes what multi-char keys are
supported/defined by the terminal? I *believe* most sequences for keys
like left, right, up, down, etc. are the same for most terminals, at
least most "xterm" terminal. So some sequences for a few common keys are
provided.

However, if the blessed module is installed, then a list of all keyboard
sequences (presumably read from terminfo) is loaded. The only potential issue with
this is that it could cause problems if the script is authored in a
different terminal than the demo. I.e. you write the script in kitty,
use vim's ctrl-v to enter sequences, and then run on gnome-terminal. If
kitty and gnome-terminal have different sequences for the same key, this
will cause unintended behavior. I'm not sure if this is actually the
case though...